Added power unit title and changed a type of one power method from float to int.

+++ src/main/java/monnef/core/block/TileMachine.java
@@ -129,7 +129,7 @@ public abstract class TileMachine extends TileMonnefCore implements IEnergyHandl
     protected void onBeforePowerSourceHandling() {
     }
 
-    public float getEnergyGeneratedThisTick() {
+    public int getEnergyGeneratedThisTick() {
         return 0;
     }
